PPM-677: Enable NBAPI in prplWrt builds
Enable with -DENABLE_NBAPI=ON only when libamxb has already been
configured.

Dependencies have to be added to the prplmesh package.
When compiling without the dependencies defined, the following error
shows:

Package prplmesh-dwpal is missing dependencies for the following libraries:
libamxb-1.so
libamxc-0.so
libamxd-1.so
libamxo-0.so
libamxp-0.so

This is an automatic lint of the package to check that the tagged
dependencies on it carry the needed libraries for the executable, which
it did not when compiling with NBAPI

Signed-off-by: pablo <pablo@odkq.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
(cherry picked from commit 52c3f0a7f7d220401fbf08d4bcf59908cf3c3d35)