prplmesh: bump to 1.5.0rc6
We still use a git sha1 to download from, because we expect that the
version will be bumped pretty soon again to a non-released version.

Refresh the patch (path to bpl_arp changed upstream).

While we're at it:
- switch to gitlab (it's no longer available on github);
- use a tarball download instead of git;
- add a hash;
- fix an incorrect <space><space><tab> indentation.

Because we download the tarball from a git sha1, we need to set the
PKG_BUILD_DIR explicitly.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
(cherry picked from commit fc848dcd516bd779330cf36e0de2b76619101559)