prpmesh: drop the package variant
It's not useful for prplMesh to be available as two different
variants: one of them is used exclusively for DWPAL-based
devices (i.e. intel_mips), and the other one is used for all the other
devices.

Remove the variant, and set the config and build options based on the
value of CONFIG_TARGET_intel_mips.

For the dependencies, use the '+TARGET_intel_mips:package' construct
to select the right packages depending on the target.

The WIRELESS_STA symbol needs to be handled differently: since it's a
symbol and not a package, we cannot use the same construct as for
packages. For that, add PRPLMESH_HAS_WIRELESS_STA to include
WIRELESS_STA by default for intel_mips.

Finally, add WIRELESS_STA and TARGET_intel_mips to PKG_CONFIG_DEPENDS,
to make sure the package is reconfigured when those symbols change.

These changes will make it easier to include prplMesh in the prpl
profile by default, simplifies the Makefile, and will also simplify
the prplMesh build scripts.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit ac2b69cf007c254dbcf275dee0d21486b41c720e)