prplmesh: bump to 1.6.0
This is a kind of backward bump because before we were on a random
commit on master. However, we prefer to work with actually released
versions, so use 1.6.0.

When gitlab creates a tarball for a tag, it normally creates a tarball
with a top-level directory named "<pkg>-<tag>-<sha1>" - except if we
request for a tarball names "<pkg>-<tag>.tar.gz", then it will use just
"<pkg>-<tag>" as the top-level directory. Since we would prefer the
latter one, make sure that we use that tarball name. Case is significant
here: the openwrt package is called "prplmesh" while the upstream
package is "prplMesh".

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
(cherry picked from commit a1a99c9acc5ac0c7907e80403978c38253271d9f)