prplmesh: uci-defaults: change the lan MAC address by default
prplMesh currently makes the assumption that the MAC address of the
LAN bridge is different than the wireless interface MACs (all of
them).

Because a big portion of prplMesh was designed with that in mind, it
will take a significant amount of time to address it properly. In the
meantime, it was decided to work around it in prplWrt by changing the
bridge MAC on devices where it can be the same as the wireless
interface MAC.

To do that, in the uci-defaults script check if the current board is
the Gl.iNet B1300. If it is, generate a new MAC address by using an
index of 10. The MAC address is generated in the same way than the
netifd scripts generate MAC addresses for hostapd. The index number 10
has been arbitrarily chosen to be big enough to allow at least two VAP
per interface.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit 549ea6967384837132124abf6591a5f675e266c2)