prplMesh: do not start if wireless is not ready
The prplMesh init script checks if the wireless is ready before
starting prplMesh.

If it's not, it logs a message but it still tries to start
prplMesh (while we know prplMesh will not work).

Change it to not start prplMesh if wireless is not ready.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit f0d2c59917d00e1b9ef32391f529d037e01eec8b)