prplmesh: add dependency on mod-amxb-ubus
When `libamxb` is selected, prplMesh is compiled with the "Northbound
API" feature enabled (`-DENABLE_NBAPI=ON`) and requires
`mod-amxb-ubus`.

This was not noticed in before, because `mod-amxb-ubus` is also
included in the `sah` profile, which is often included when `libamxb`
is included.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit fb8b4a6b0f57b6a86a065eb7fc73620683a6fb8d)