station-sniffer: updated service files
Currently, the station-sniffer uses the no-longer viable `/etc/config/wireless`
file to find the `phy` for a specific frequency (hardcoded at 5 GHz). The removal
of `wireless` in favor of the ambiorix data, the model renders the `station-sniffer`
The service is defunct, so it needs to be updated to utilize the data model.

This commit makes the station-sniffer service usable again by adding support
for the ambiorix data model through the `ba-cli` command, along with new support
for a `uci` config file to specify the frequency to start the monitor mode interface on.

Signed-off-by: Benjamin Carlson <b.carlson@cablelabs.com>
(cherry picked from commit d0e3dacbfcd2730b249b2d3870893adf27e943b6)