prplmesh: change restart function of the init script
At the current state the function is calling stop + start, which does
not allow the processmonitor to correctly restart prplMesh because
its subprocess is being killed when the processmonitor entry is removed
while executing the restart function.

The start_function will never be reached in this case, because the processmonitor entry that is
responsible for restarting is deleted.

Therefore, we call the restart function from the prplmesh_utils directly, which does not remove the
prplMesh ProcessMonitor entries that are already present in the datamodel.

References: PPM-3129
Signed-off-by: Denys Stolbov <d.stolbov@inango-systems.com>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org> [commit facelift]
(cherry picked from commit 2682f9118d8a7535a7f64f15a18bd6cb2d334fde)