prplMesh: patch prplMesh ACL and configure proxy path
PrplMesh's ACL only defines permissions for the @CONTROLLER_ROOT_DM@ path, which is the controller configuration datamodel.
In order for the TR-181 dataelements DM to be available over the REST API, additional permissions need to be set on the @DATAELEMENTS_ROOT_DM@ path.

In order to correctly map the dataelements DM to Device.WiFi.DataElements, an extra mapper ODL is added.

These changes are provided as a patch to the prplMesh v5.0 release, and will be merged in prplMesh itself in the next version release.

See: PPM-3256
See: PPM-3311
Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>