prplmesh: install prplmesh ACL file during build.
The `prplmesh.json` ACL file is not currently installed on the prplOS image due to
missing installation steps in the `prplmesh/Makefile`. This file is essential for
the proper access control of the prplMesh WiFi controller `X_PRPL-ORG_WiFiController.` and `Device.WiFi.DataElements`.

- Adding commands to create the `/etc/acl/admin` directory during installation.
- Copying the `prplmesh.json` file to the `/etc/acl/admin` directory.

These changes ensure that the prplMesh ACL is correctly included in the build,
fixing the installation issue mentioned in PPM-3007.

Closes: PPM-3007, PPM-3256

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>