prplmesh: install prplmesh proxymapping ODL
PrplMesh' proxymapping ODL is not currently installed on the prplOS image due to missing installation steps in the `prplmesh/Makefile`.
This file takes care of the correct controller datamodel mapping, and maps "X_PRPLWARE-COM_WiFiController." to "Device.WiFi.DataElements.".
These changes make sure it is installed in the tr181-device/extensions folder.

Closes: PPM-3311

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>